Fix bug causing writes to $2006 to set flags incorrectly in _Address struct (#40)
The register (reg) value in the union for the _Address class was a 16-bit values and the individual flags were 8-bit values. A union is the size of its largest value, and this type mismatch was causing incorrect values to be stored in coarse_y, nt_select and fine_y. This change updates all of the values in the bitfield struct to be 16-bit values to prevent this. Unit tests have been added to cover this test case.

+++ purenes/ppu.py
@@ -167,10 +167,10 @@ class _Address(ctypes.Union):
             "_PPUADDRESS",
             (ctypes.LittleEndianStructure,),
             {"_fields_": [
-                ("coarse_x",  ctypes.c_uint8, 5),
-                ("coarse_y",  ctypes.c_uint8, 5),
-                ("nt_select", ctypes.c_uint8, 2),
-                ("fine_y",    ctypes.c_uint8, 3),
+                ("coarse_x",  ctypes.c_uint16, 5),
+                ("coarse_y",  ctypes.c_uint16, 5),
+                ("nt_select", ctypes.c_uint16, 2),
+                ("fine_y",    ctypes.c_uint16, 3),
             ]}
         )),
         ("reg", ctypes.c_uint16)]
